+++ srcs/sorts/comparators/objects/sort_time_objects_cmp.c
@@ -6,13 +6,13 @@
 /*   By: tmaluh <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/07/31 16:29:03 by tmaluh            #+#    #+#             */
-/*   Updated: 2019/11/30 18:37:44 by tmaluh           ###   ########.fr       */
+/*   Updated: 2019/12/08 16:38:20 by tmaluh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ls.h"
 
-int32_t	sort_time_objects_cmp(const void *a, const void *b)
+int64_t	sort_time_objects_cmp(const void *a, const void *b)
 {
 	const struct s_stat	*st_a = ((const struct s_object*)a)->st;
 	const struct s_stat	*st_b = ((const struct s_object*)b)->st;
@@ -24,5 +24,5 @@ int32_t	sort_time_objects_cmp(const void *a, const void *b)
 		cmp = st_a->m_time - st_b->m_time;
 	if (!cmp)
 		return (sort_ascii_objects_cmp(a, b));
-	return ((int32_t)(IS_BIT(g_flags, BIT_R_SORT_REV) ? cmp : -cmp));
+	return ((IS_BIT(g_flags, BIT_R_SORT_REV) ? cmp : -cmp));
 }
